[Feat]날씨화면 수평 스크롤 구현
